+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 ## 1.7.0
 
-* File selection with support Android 13.
+* Other Device option.
 
 ## 1.6.0